+++ ChangeLog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ## 1.5
+- FIX: Duplication of breadcrumb bar on older Dolibarr versions *09/02/2025* - 1.5.1
 - FIX: Compatibility with DARK THEMES *08/02/2025* - 1.5.0
 - NEW: Spanish translation *08/02/2025* - 1.5.0
 - FIX: Correctly translated the name of module on Settings grid of modules *08/02/2025* - 1.5.0